+++ Okta.AspNetCore/CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 Running changelog of releases since `3.2.0`
 
+## v4.4.0
+
+### Features
+
+- Add support for `prompt` and `enroll_amr_values`
+
 ## v4.3.0
 
 ### Features
 
-- Add support for AcrValues
+- Add support for `acr_values`
 
 ## v4.2.1